requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements/base.txt --resolver=backtracking pyproject.toml
#
aiohttp==3.8.4
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.1
    # via aiohttp
appdirs==1.4.4
    # via wandb
async-timeout==4.0.2
    # via aiohttp
attrs==22.2.0
    # via aiohttp
certifi==2022.12.7
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.0.1
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   typer
    #   wandb
colorama==0.4.6
    # via typer
commonmark==0.9.1
    # via rich
contourpy==1.0.7
    # via matplotlib
cycler==0.11.0
    # via matplotlib
datasets==2.10.1
    # via vqa (pyproject.toml)
dill==0.3.6
    # via
    #   datasets
    #   multiprocess
docker-pycreds==0.4.0
    # via wandb
filelock==3.9.0
    # via huggingface-hub
fonttools==4.38.0
    # via matplotlib
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2023.3.0
    # via datasets
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via wandb
huggingface-hub==0.12.1
    # via datasets
idna==3.4
    # via
    #   requests
    #   yarl
joblib==1.2.0
    # via scikit-learn
kiwisolver==1.4.4
    # via matplotlib
matplotlib==3.7.1
    # via vqa (pyproject.toml)
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.14
    # via datasets
numpy==1.24.2
    # via
    #   contourpy
    #   datasets
    #   matplotlib
    #   pandas
    #   pyarrow
    #   scikit-learn
    #   scipy
    #   torchvision
nvidia-cublas-cu11==11.10.3.66
    # via
    #   nvidia-cudnn-cu11
    #   torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via torch
nvidia-cuda-runtime-cu11==11.7.99
    # via torch
nvidia-cudnn-cu11==8.5.0.96
    # via torch
packaging==23.0
    # via
    #   datasets
    #   huggingface-hub
    #   matplotlib
pandas==1.5.3
    # via datasets
pathtools==0.1.2
    # via wandb
pillow==9.4.0
    # via
    #   matplotlib
    #   torchvision
protobuf==4.22.0
    # via wandb
psutil==5.9.4
    # via wandb
pyarrow==11.0.0
    # via datasets
pygments==2.14.0
    # via rich
pyparsing==3.0.9
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.0
    # via vqa (pyproject.toml)
pytz==2022.7.1
    # via pandas
pyyaml==6.0
    # via
    #   datasets
    #   huggingface-hub
    #   wandb
requests==2.28.2
    # via
    #   datasets
    #   fsspec
    #   huggingface-hub
    #   responses
    #   torchvision
    #   wandb
responses==0.18.0
    # via datasets
rich==12.6.0
    # via typer
scikit-learn==1.2.1
    # via vqa (pyproject.toml)
scipy==1.10.1
    # via scikit-learn
sentry-sdk==1.16.0
    # via wandb
setproctitle==1.3.2
    # via wandb
shellingham==1.5.0.post1
    # via typer
six==1.16.0
    # via
    #   docker-pycreds
    #   python-dateutil
smmap==5.0.0
    # via gitdb
threadpoolctl==3.1.0
    # via scikit-learn
torch==1.13.1
    # via
    #   torchvision
    #   vqa (pyproject.toml)
torchvision==0.14.1
    # via vqa (pyproject.toml)
tqdm==4.65.0
    # via
    #   datasets
    #   huggingface-hub
typer[all]==0.7.0
    # via vqa (pyproject.toml)
typing-extensions==4.5.0
    # via
    #   huggingface-hub
    #   torch
    #   torchvision
urllib3==1.26.14
    # via
    #   requests
    #   responses
    #   sentry-sdk
wandb==0.13.10
    # via vqa (pyproject.toml)
wheel==0.38.4
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-runtime-cu11
xxhash==3.2.0
    # via datasets
yarl==1.8.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
